FAC-572: Refactor - Adapt Accept Assessment Invitations Notification With New Architecture
