+++ flickit-assessment-common/src/main/resources/i18n/core/messages_en.properties
@@ -264,6 +264,7 @@ subject-default-insight=<p>One key aspect of this assessment is evaluating the m
 notification-title.grant-assessment-user-role=Assessment Access Granted
 notification-title.create-assessment=New Assessment on Your Kit
 notification-title.completed-assessment=Assessment Completed
+notification-title.accept-assessment-invitation=Invitation Accepted
 
 add-assessment-analysis-input-file.assessmentId.notNull='assessmentId' may not be null
 add-assessment-analysis-input-file.inputFile.notNull='inputFile' may not be null
