+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/domain/notification/AcceptAssessmentInvitationNotificationCmd.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.core.application.service.assessmentinvite.notification;
+package org.flickit.assessment.core.application.domain.notification;
 
 import org.flickit.assessment.common.application.domain.notification.NotificationCmd;
 import org.flickit.assessment.core.application.domain.AssessmentUserRole;
@@ -6,11 +6,12 @@ import org.flickit.assessment.core.application.domain.AssessmentUserRole;
 import java.util.List;
 import java.util.UUID;
 
-public record AcceptAssessmentInvitationNotificationCmd(List<NotificationCmdItem> notificationCmdItems) implements NotificationCmd {
+public record AcceptAssessmentInvitationNotificationCmd(
+    List<NotificationCmdItem> notificationCmdItems) implements NotificationCmd {
 
     public record NotificationCmdItem(UUID targetUserId,
                                       UUID assessmentId,
                                       UUID inviteeId,
-                                      AssessmentUserRole assessmentUserRole) {
+                                      AssessmentUserRole assessmentUserRole) implements NotificationCmd {
     }
 }
