+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/service/assessmentinvite/notification/AcceptAssessmentInvitationNotificationCreator.java
@@ -2,10 +2,12 @@ package org.flickit.assessment.core.application.service.assessmentinvite.notific
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.flickit.assessment.common.application.MessageBundle;
 import org.flickit.assessment.common.application.domain.notification.NotificationCreator;
 import org.flickit.assessment.common.application.domain.notification.NotificationEnvelope;
 import org.flickit.assessment.core.application.domain.Assessment;
 import org.flickit.assessment.core.application.domain.User;
+import org.flickit.assessment.core.application.domain.notification.AcceptAssessmentInvitationNotificationCmd;
 import org.flickit.assessment.core.application.port.out.assessment.GetAssessmentPort;
 import org.flickit.assessment.core.application.port.out.user.LoadUserPort;
 import org.flickit.assessment.core.application.service.assessmentinvite.notification.AcceptAssessmentInvitationNotificationPayload.*;
@@ -14,6 +16,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
+
+import static org.flickit.assessment.core.common.ErrorMessageKey.NOTIFICATION_TITLE_ACCEPT_ASSESSMENT_INVITATION;
 
 @Slf4j
 @Component
@@ -35,15 +39,20 @@ public class AcceptAssessmentInvitationNotificationCreator
     public NotificationEnvelope creator(AcceptAssessmentInvitationNotificationCmd.NotificationCmdItem cmd) {
         Optional<Assessment> assessment = getAssessmentPort.getAssessmentById(cmd.assessmentId());
         Optional<User> user = loadUserPort.loadById(cmd.inviteeId());
-        if (assessment.isEmpty() || user.isEmpty()) {
+        var targetUser = loadUserPort.loadById(cmd.targetUserId())
+            .map(x -> new NotificationEnvelope.User(x.getId(), x.getEmail()));
+        if (assessment.isEmpty() || user.isEmpty() || targetUser.isEmpty()) {
             log.warn("assessment or user not found");
             return null;
         }
 
-        return new NotificationEnvelope(cmd.targetUserId(), new AcceptAssessmentInvitationNotificationPayload(
-            new AssessmentModel(assessment.get().getId(), assessment.get().getTitle()),
-            new InviteeModel(user.get().getId(), user.get().getDisplayName()),
-            new RoleModel(cmd.assessmentUserRole().getTitle())));
+        var title = MessageBundle.message(NOTIFICATION_TITLE_ACCEPT_ASSESSMENT_INVITATION);
+        var payload = new AcceptAssessmentInvitationNotificationPayload(
+            new AssessmentModel(cmd.assessmentId(), assessment.get().getTitle()),
+            new InviteeModel(cmd.inviteeId(), user.get().getDisplayName()),
+            new RoleModel(cmd.assessmentUserRole().getTitle()));
+
+        return new NotificationEnvelope(targetUser.get(), title, payload);
     }
 
     @Override
